patches: cleanup to separate 600 and 710 patches, and add 710 reflect+refract res patching
