+++ source-code/source/plugins/Patches/PatchApplier.h
@@ -1,0 +1,7 @@
+#pragma once
+
+// PatchApplier provides the stuff for implementing patches for the specific game version
+class PatchApplier {
+public:
+	virtual void ApplyPatches() {};
+};
